QuowWE: Reindent code

+++ d3web/d3web-CaseRepository/src/test/java/de/d3web/caserepository/test/CaseRepositoryPersistenceTest.java
@@ -1,4 +1,5 @@
 package de.d3web.caserepository.test;
+
 import java.io.File;
 import java.io.IOException;
 
@@ -13,11 +14,10 @@ import de.d3web.core.records.io.MultipleXMLCaseRepositoryHandler;
 import de.d3web.core.records.io.SingleXMLCaseRepositoryHandler;
 import de.d3web.plugin.test.InitPluginManager;
 
-
 public class CaseRepositoryPersistenceTest extends TestCase {
-	
+
 	KnowledgeBase kb;
-	
+
 	public CaseRepositoryPersistenceTest(String arg0) {
 		super(arg0);
 	}
@@ -25,11 +25,11 @@ public class CaseRepositoryPersistenceTest extends TestCase {
 	public static void main(String[] args) {
 		TestRunner.run(CaseRepositoryPersistenceTest.suite());
 	}
-	
+
 	public static Test suite() {
 		return new TestSuite(CaseRepositoryPersistenceTest.class);
 	}
-	
+
 	protected void setUp() {
 		try {
 			InitPluginManager.init();
@@ -37,11 +37,12 @@ public class CaseRepositoryPersistenceTest extends TestCase {
 		catch (IOException e1) {
 			assertTrue("Error initialising plugin framework", false);
 		}
-		//create tests folders
+		// create tests folders
 		new File("target/tests/multi").mkdirs();
 		try {
 			loadKnowledgeBase();
-		} catch (IOException e2) {
+		}
+		catch (IOException e2) {
 			assertTrue("Error while loading KnowledgeBase", false);
 		}
 	}
@@ -51,68 +52,80 @@ public class CaseRepositoryPersistenceTest extends TestCase {
 		String kbPath = System.getProperty("user.dir") + "/src/test/resources/carDiagnosis.jar";
 		kb = pm.load(new File(kbPath));
 	}
-	
+
 	public void testSingleXMLLoading() throws Exception {
-		File caseFile = new File(System.getProperty("user.dir") + "/src/test/resources/carDiagnosis_cases.xml");
+		File caseFile = new File(System.getProperty("user.dir")
+				+ "/src/test/resources/carDiagnosis_cases.xml");
 		CaseRepository repository = SingleXMLCaseRepositoryHandler.getInstance().load(kb, caseFile);
 		assertNotNull("The loaded repository is null.", repository);
-		assertNotNull("CaseObject with ID \"855859\" wasn't found.", repository.getCaseObjectById("8558590"));
-		assertNotNull("CaseObject with ID \"9979645\" wasn't found.", repository.getCaseObjectById("9979645"));
-		assertNotNull("CaseObject with ID \"10981489\" wasn't found.", repository.getCaseObjectById("10981489"));
+		assertNotNull("CaseObject with ID \"855859\" wasn't found.",
+				repository.getCaseObjectById("8558590"));
+		assertNotNull("CaseObject with ID \"9979645\" wasn't found.",
+				repository.getCaseObjectById("9979645"));
+		assertNotNull("CaseObject with ID \"10981489\" wasn't found.",
+				repository.getCaseObjectById("10981489"));
 	}
-	
+
 	public void testSingleXMLSaving() throws Exception {
 		// Load the old file / repository
-		File oldFile = new File(System.getProperty("user.dir") + "/src/test/resources/carDiagnosis_cases.xml");
+		File oldFile = new File(System.getProperty("user.dir")
+				+ "/src/test/resources/carDiagnosis_cases.xml");
 		CaseRepository repository = SingleXMLCaseRepositoryHandler.getInstance().load(kb, oldFile);
-		
+
 		// Save the new file / repository
 		File newFile = new File("target/tests/carDiagnosis_cases_new.xml");
 		SingleXMLCaseRepositoryHandler.getInstance().save(repository, newFile);
-		
+
 		// Load the new file / repository
 		repository = SingleXMLCaseRepositoryHandler.getInstance().load(kb, newFile);
-		
+
 		assertNotNull("The loaded repository is null.", repository);
-		assertNotNull("CaseObject with ID \"855859\" wasn't found.", repository.getCaseObjectById("8558590"));
-		assertNotNull("CaseObject with ID \"9979645\" wasn't found.", repository.getCaseObjectById("9979645"));
-		assertNotNull("CaseObject with ID \"10981489\" wasn't found.", repository.getCaseObjectById("10981489"));
-		
+		assertNotNull("CaseObject with ID \"855859\" wasn't found.",
+				repository.getCaseObjectById("8558590"));
+		assertNotNull("CaseObject with ID \"9979645\" wasn't found.",
+				repository.getCaseObjectById("9979645"));
+		assertNotNull("CaseObject with ID \"10981489\" wasn't found.",
+				repository.getCaseObjectById("10981489"));
+
 		newFile.delete();
 	}
-	
+
 	public void testMultipleXMLSaving() throws Exception {
 		// Load the old file / repository
-		File oldFile = new File(System.getProperty("user.dir") + "/src/test/resources/carDiagnosis_cases.xml");
+		File oldFile = new File(System.getProperty("user.dir")
+				+ "/src/test/resources/carDiagnosis_cases.xml");
 		CaseRepository repository = SingleXMLCaseRepositoryHandler.getInstance().load(kb, oldFile);
-		
+
 		// Save the repositories to multiple files
 		File directory = new File("target/tests/multi");
 		MultipleXMLCaseRepositoryHandler.getInstance().save(repository, directory);
-		
+
 		// Test the number of files in the directory
 		int fileCounter = 0;
 		for (File f : directory.listFiles()) {
-			if (f.getName().endsWith(".xml"))
-				fileCounter++;
+			if (f.getName().endsWith(".xml")) fileCounter++;
 		}
 		assertEquals("There should be three XML Files.", 3, fileCounter);
 	}
-	
+
 	public void testMultipleXMLLoading() throws Exception {
-		// Load the old files / repository (the files are created in the test before)
+		// Load the old files / repository (the files are created in the test
+		// before)
 		File directory = new File("target/tests/multi");
-		CaseRepository repository = MultipleXMLCaseRepositoryHandler.getInstance().load(kb, directory);
-		
+		CaseRepository repository = MultipleXMLCaseRepositoryHandler.getInstance().load(kb,
+				directory);
+
 		assertNotNull("The loaded repository is null.", repository);
-		assertNotNull("CaseObject with ID \"855859\" wasn't found.", repository.getCaseObjectById("8558590"));
-		assertNotNull("CaseObject with ID \"9979645\" wasn't found.", repository.getCaseObjectById("9979645"));
-		assertNotNull("CaseObject with ID \"10981489\" wasn't found.", repository.getCaseObjectById("10981489"));
-		
+		assertNotNull("CaseObject with ID \"855859\" wasn't found.",
+				repository.getCaseObjectById("8558590"));
+		assertNotNull("CaseObject with ID \"9979645\" wasn't found.",
+				repository.getCaseObjectById("9979645"));
+		assertNotNull("CaseObject with ID \"10981489\" wasn't found.",
+				repository.getCaseObjectById("10981489"));
+
 		// Delete the files in the directory
 		for (File f : directory.listFiles()) {
-			if (f.getName().endsWith(".xml"))
-				f.delete();
+			if (f.getName().endsWith(".xml")) f.delete();
 		}
 	}
